Debug snesbrr

+++ tools/snesbrr/base/basemath.h
@@ -63,7 +63,7 @@ namespace base
         template <uint N, typename T>
         inline static T clamp(T x)
         {
-            const T low  = static_cast<T>(-1) << (N - 1);
+            const T low = static_cast<T>(static_cast<unsigned int>(-1) << (N - 1));
             const T high = (1 << (N - 1)) - 1;
 
             if (x > high)
